Implement GlyLong for use in ClassBinder
When using the Enroute.iOS Xamarin library, an event would arrive to be handled in the eventsOccured method of the Glympse.EnRoute.iOS.ListenerWrapper class.

This would cause an Exception to be thrown in our application. We obtained the Enroute.iOS library and referenced the source rather than the library DLL. Doing this, we discovered the Exception was coming from ClassBinder, which did not know how to bind the type arriving, which was a GlyLong.

The “param1” parameter of type GlyCommon was the culprit: this was a GlyLong. Referencing what we found at https://developer.glympse.com/docs/core/client-sdk/reference/ios/Long, we created an implementation of GlyLong within ApiDefinition.cs and added a clause to Classbinder.

This appears to solve the problem, we now receive the event properly, with no Exception.